refactor(automation): consolidate agent config into single module
Merge claude_models, claude_timeouts, and session_naming into agent_config.py to eliminate cross-imports and clarify the agent configuration domain. Reduces module count from 4 to 2 while keeping subprocess logic in claude_invoke.py separate.

Closes #1441

Implemented-By: claude-opus-4-8
